Query config factory has query modules.
Query config factory was growing too fast.
Added modules in order to distribute the responsability of building each query.
